milestone 1/requirements.txt
streamlit
pyjwt
bcrypt
pyngrok